+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/monitor/AutocraftingMonitor.java
@@ -6,4 +6,6 @@ interface AutocraftingMonitor extends TaskStatusProvider {
     void addWatcher(AutocraftingMonitorWatcher watcher);
 
     void removeWatcher(AutocraftingMonitorWatcher watcher);
+
+    boolean isAutocraftingMonitorActive();
 }