+++ CHANGELOG.md
@@ -10,6 +10,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 -   Autocrafting Monitor
+-   Wireless Autocrafting Monitor
+-   Creative Wireless Autocrafting Monitor
 
 ### Changed